Refactor Newfeed to container NotificationView, Main Component Notification, subComponent NotificationList and NotificationItem
