+++ assets/app/js/config/routes.jsx
@@ -3,7 +3,7 @@ import {Route, IndexRoute, IndexRedirect} from 'react-router'
 import App from '../containers/App.jsx'
 import Project from '../containers/Project.jsx'
 import _404 from '../components/_404.jsx'
-import Notifications from '../components/Notifications.jsx'
+import NotificationView from '../containers/NotificationView.jsx'
 import SearchResults from '../components/SearchResults.jsx'
 import Dashboard from '../components/Dashboard.jsx'
 import * as AppConstants from '../AppConstants';
@@ -18,7 +18,7 @@ export default (
       // <Route path={AppConstants.PATH.newsfeed} />
       // <Route path={AppConstants.PATH.settings} />
     </Route>
-    <Route path='notifications' component={Notifications}/>
+    <Route path='notifications' component={NotificationView}/>
     <Route path='search' component={SearchResults}/>
     <Route path='dashboard' component={Dashboard}/>
     <Route path='*' component={_404}/>
